--- a/diceSimulator.c
+++ b/diceSimulator.c
@@ -8,6 +8,9 @@
 
 #define ARRAY_SIZE DIE1_NUM_SIDES + DIE2_NUM_SIDES - 1
 
+unsigned int pos_ceil(double num);
+void progress_bar(double complete, double total);
+
 int main(void) {
     unsigned int i, results[ARRAY_SIZE];
     
@@ -26,6 +29,8 @@ int main(void) {
 
         total = dice1 + dice2;
         results[total] += 1;
+
+        progress_bar(i+1, SAMPLE_SIZE);
     }
 
     /* print out the results */
@@ -43,4 +48,40 @@ int main(void) {
     system("Rscript ./plot.r &>/dev/null");
 
     return 0;
+}
+
+/**
+ * Print out a progress bar.
+ * Only updates when it needs to.
+ */
+void progress_bar(double complete, double total) {
+    unsigned int i, perc;
+    static unsigned int old_perc = 0;
+
+    perc = pos_ceil(100*(complete/total));
+
+    if (perc == old_perc) return;
+    old_perc = perc;
+
+    printf("\r%u%% ", perc);
+    for (i = 0; i < perc/2; i++)
+        printf("#");
+
+    if (perc == 100)
+        printf("\n");
+
+    fflush(stdout);
+}
+
+/**
+ * Round up.
+ * Only works for pos. numbers.
+ */
+unsigned int pos_ceil(double num) {
+    int inum = (int)num;
+
+    if (num == (float)inum)
+        return inum;
+
+    return ((unsigned int) inum + 1);
 }