Don't take this too harshly. I just had some free time and went through your code line by line.

You can see my reasoning for each change by viewing the full commit messages by clicking on each commit below (or by using `git log`).

The speed improvements aren't much, but they're there! (about 0.4 seconds faster for 1000 trials).

```
1000 Trials of Sample Size 1000

===== BEFORE =====
Total = 6.385s
Avg.  = 0.006s

===== AFTER =======
Total = 5.986s
Avg.  = 0.006s
```
(btw, I timed each program using my `avgtime` command)